Rename file {specialist_document, section}_repository
