+++ app/repositories/repository_registry.rb
@@ -1,4 +1,4 @@
-require "specialist_document_repository"
+require "section_repository"
 require "manual_repository"
 require "section_edition"
 require "marshallers/document_association_marshaller"
@@ -7,7 +7,7 @@ require "manual_publish_task"
 require "manual_with_publish_tasks"
 require "manual"
 require "manual_record"
-require "specialist_document_repository"
+
 
 class RepositoryRegistry
   def self.create
