+++ spec/repositories/section_repository_spec.rb
